fix(neovim): disable fterm in terminal mode
Causes which-key to watch for the leader key because a keymap exists in
that mode.

The consequences are that whenever you press the leader key, it does not
appear in the terminal until you press the next key which in this case
means when I press the Space key, it looks like the terminal didn't
register it and is super annoying.

Pressing ESC twice to exit terminal mode into normal mode is enough of a
solution to then use whatever keymaps I have available.