fix(neovim): specify plugin names + add luvit-meta
For whatever reason, after a nixpkgs channel update, lazy could not
longer find luasnip and catppuccin and also complained that the plugin
`luvit-meta` could not be find.

Uncertain what causes this but likely a new feature/bug in Lazy. Adding
`name` attribute seems to force lazy.nvim to look in the right folder
and we were indeed missing the `luvit-meta` plugin.